requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
chromadb==0.5.5
PyMuPDF==1.24.9
python-dotenv==1.0.1
langchain==0.3.3
langchain-community==0.3.2
openai==1.51.2
